fix(web): spawn worktree child against launching parent on mobile/web AutoRun
Mobile/web AutoRun's "Dispatch to a separate worktree" was relying on the
chokidar watcher in useWorktreeHandlers to discover the new on-disk
worktree and attach it to a parent. The watcher's find-first heuristic
attached the child to whichever sibling agent's worktreeConfig.basePath
matched first, producing the wrong-parent attachment reported in PR #946.

Extract spawnWorktreeAgentAndDispatch into a shared util and call it
from the remote configureAutoRun handler, mirroring desktop's
useAutoRunHandlers parent resolution. The remote path now spawns the
child session explicitly with the correct parentSessionId before
startBatchRun runs, and forwards worktreeTarget so startBatchRun skips
its redundant setupWorktree.

Also closes Greptile P1: AutoRunWorktreeSection now emits enabled-invalid
when createPROnCompletion is true but baseBranch is still empty (race
window between loadBranches resolution and the user toggling PR
creation).
